only tests leftttttt

+++ signals.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <signal.h>
 #include <sys/wait.h>
+#include <unistd.h>
 #include "signals.h"
 #include "Commands.h"
 
@@ -8,35 +9,40 @@ using namespace std;
 
 void ctrlZHandler(int sig_num) {
     SmallShell &smash = SmallShell::getInstance();
-    cout << "smash: got ctrl-Z" << endl;
+    string print = "smash: got ctrl-Z\n";
+    write_to(print, smash.getOut());
     if (smash.getPidInFG() != -1) {
         kill(smash.getPidInFG(), SIGTSTP);
         if (smash.getJob())
             smash.jobs->addJob(smash.getJob());
         else
             smash.jobs->addJob(smash.getCommand(), smash.getPidInFG(), 0);
-        cout << "smash: process " + to_string(smash.getPidInFG()) + " was stopped" << endl;
-        smash.setCommand(nullptr);
-        smash.setJob(nullptr);
-        smash.setPidInFG(-1);
+        print = "smash: process " + to_string(smash.getPidInFG()) + " was stopped\n";
+        write_to(print, smash.getOut());
+        smash.setNulls();
     }
 }
 
 void ctrlCHandler(int sig_num) {
     SmallShell &smash = SmallShell::getInstance();
-    cout << "smash: got ctrl-C" << endl;
+    string print = "smash: got ctrl-C\n";
+    write_to(print, smash.getOut());
     if (smash.getPidInFG() != -1) {
         kill(smash.getPidInFG(), SIGKILL);
-        cout << "smash: process " + to_string(smash.getPidInFG()) + " was killed" << endl;
+        print = "smash: process " + to_string(smash.getPidInFG()) + " was killed\n";
+        write_to(print, smash.getOut());
     }
-    smash.setCommand(nullptr);
-    smash.setJob(nullptr);
-    smash.setPidInFG(-1);
+    smash.setNulls();
 }
 
 void alarmHandler(int sig_num) {
     SmallShell &smash = SmallShell::getInstance();
-    cout << "smash got an alarm" << endl;
-    kill(smash.getPidInFG(), SIGKILL);
+    string print = "smash got an alarm\n";
+    write_to(print, smash.getOut());
+    if (smash.getPidInFG() != -1) {
+        kill(smash.getPidInFG(), SIGKILL);
+        cout << "smash: " << smash.getCommand()->getJob() << " timed out!" << endl;
+    }
+    smash.setNulls();
 }
